Create Litz
Hello from LITZ and LITBITZ. Only line not certian what we need to do is line 8. Are we okay to leave line 8 as is? THANK YOU ALL VERY MUCH AND HAPPY NEW YEARS!

our coin is bought in this manor:
Send a minimum of 3 ADA and you get back 1.38 ADA + Random LITZ from 10k to 1M.
Send multiples of 1 ADA to get back more LITZ